refactor(claude-tests): split CLI command tests out of tests_core for large-file gate
tests_core.rs was over the large-file governance threshold once perf
baseline scripts were added to the repo budget. Move Claude CLI command
construction tests into a dedicated tests_command module so the gate
passes without weakening test coverage.

- Register tests_command module in claude.rs alongside the existing test modules
- Extract Claude CLI command construction cases into src-tauri/src/engine/claude/tests_command.rs
- Trim tests_core.rs to 2368 lines so check:large-files:gate reports found=0
