+++ src-tauri/src/engine/claude.rs
@@ -2111,6 +2111,9 @@ impl ClaudeSession {
 #[path = "claude/tests_context_usage.rs"]
 mod tests_context_usage;
 #[cfg(test)]
+#[path = "claude/tests_command.rs"]
+mod tests_command;
+#[cfg(test)]
 #[path = "claude/tests_core.rs"]
 mod tests_core;
 #[cfg(test)]
